Clean up the LayoutItem type (#8235)
